R&D (ingame-terminal): Concluded researching implementation opts.
**Some important things to remember I've learned along my journey:**

ImGui is not Object Oriented Programming qualified so there are
no window classes or objects to fiddle with. This feature is
actually a little confusing because while the design is linear,
there are some features that could better benefit from having
object inheritance for ease-of-use such as element sizing
and accessor functions. As it currently stands all the accessors
are written in a way such that you can only grab information
from within the block definition of the objects you're trying to
refference. After that, you loose the ability to create new
refferences to those values. This isn't super constrictive, but
it means that if you want to embed responsive elements into an
interface, you have to create forward definitions for the values
you will be querying ahead of the elements you're querying them from
*and* ahead of the elements that will be using those values.
It really just looks like a giant glob of spaghetti code.

Element specific styling can be implemented, but it's extremely
tedious. Any change to the styling that's done in a render loop
and is targeted at only one element must be applied via a
push & pop procedure that must wrap around the end definitions
of the frames in which the styling applies to. Also, said
methods for invoking that styling, ie. `PushStyleVal` and
`PushStyleColor` respectively; only accept one feature definition
at a time. For example, if I wanted to change the border and
frame's background color at the same time I'd need to
```c++
	ImGui::PushStyleColor(ImGuiCol_Border, IM_COL32(100,100,100,255));
	ImGui::PushStyleColor(ImGuiCol_FrameBg, IM_COL32(90,10,90,255));
	// put elements with selected styling below.

	// and then have a
	ImGui::PopStyleColor();
	// for however many pushes you've invoked.
```

There was also one API inconsistency demonstrated in the
`ImGui::InputText...()` widgets, geared at responsive design.
at a time. For example, if I wanted to change the border and
frame's background color at the same time I'd need to
```c++
	ImGui::PushStyleColor(ImGuiCol_Border, IM_COL32(100,100,100,255));
	ImGui::PushStyleColor(ImGuiCol_FrameBg, IM_COL32(90,10,90,255));
	// put elements with selected styling below.

	// and then have a
	ImGui::PopStyleColor();
	// for however many pushes you've invoked.
```

There was also one API inconsistency demonstrated in the
`ImGui::InputText...()` widgets, geared at responsive design.
For some reason as I've observed this is the only element to
feature a callback function argument, and said callback
function argument can handle a wide array of interrrupts
such as tabs for autocompletion, up and down for output history,
etc. for some reason, the API insists on having the widget draw
function return true on *return / enter* key press however. I spent
four hours trying to figure that nonesense out.