feat (ingame-terminal): Created OOP scaffold, preview is working.
As of right now, I've established the Object Oriented interface
for further extensions. The `TerminalBase` class establishes
all basic features a terminal needs. `BasicTerminal` is a simple
basic terminal that implements all the necessary features.

I've created a UI container within the sandbox directory for
ingame UI elements. Anything in the 'Engine' is for general use,
the UI elements found within the sandbox are specifically for
phoenix. The kind of terminal @apachano was suggesting reminded
me of the minecraft chat window that doubled as a command executor
so I based my implementation on that.

Existing notable features:
 * Output from the terminal is displayed above the edit field in
   both the base class and the chat UI implement.
 * Hitting return sends the contents of the edit field to a stream.
   As of right now it's a vector string stream, but I'm looking
   to change that to a `std::stringstream` object for better
   versaitlity and standard library compliance / ease-of-use.