feat (ingame-terminal): Added chat window styling!
Finally the chat window is starting to look like
Minecraft's! I also consolidated the API for
the terminal widget a little bit. By no means
does it adhear to the ImGui conventions but it
uses them under the hood now without any hacks
so it should be seemless to extend off now.

I'm going to try and start working on the
per-terminal stream outputs now. Until next commit!