refactor (ingame-terminal): Rewrote ImGuiHelpers with imgui_internal.
Getting ready for complete change over to imgui_internal code coverage
since I finally figured out how to use it. Got rid of useless
TerminalBase and created ImOOGuiWindowBase for implementing
Object Oriented ImGui Windows / UI elements.

Also I took the liberty of removing those pesky virtual methods
because making them virtual only adds extra runtime computations
to check whether a pointer of that target function belongs to
the ancestor which initially derived it or a decendant that
derived it to check which implementation to call. Damn I learn
quick. So really there's no benefit to having those be
virtual when initially I thought things had to be virtual for
you to override them, but of course that's why inheritance
classifiers are a thing...