refactor(ingame-terminal): Changed callback format.
Callbacks are now registered on a per-terminal basis
and can be specified at runtime. More complex features may
be added later as needed, but for now this change was implemented
to reflect the standard library way of handling callbacks on a
per class basis that's more open. You no longer have to subclass
to change the behavior of a terminal's callbacks. This is also
makes it easier to change the terminals's callback types to
asynchronous so they can use the cpu instead of the gpu / render
pipeline later on.