refactor(ingame-terminal): Changed callback loop to ranged-for.
This is a *potential* performance gain; but most importantly a
undefined behavior prevention. Looping over the elements of an
array and modifying them via their respective itterator lists
is undefined behavior. While the data we're grabbing is currently
read only, it's good to be able to change that in the future
without unwarranted consciquences.

Thanks Blacklegi aka. @DontBelieveMe for the tip.