doc(filesystem): Added external credit to StackOverflow code.
Which I feel inclined to say is not technically licensed / considered
Open Source and therefore could result in legal problems down the road.
I do this from time to time in my own projects but that's only because
I seek to make no money from them. If the project does develop
some income, depending on how large, this may eventually become a
problem.