fix(feat-command): Forgot to remove `move` in post function.
So, here's something I forgot to mention: we don't actually need
move semantics for the run function at all. It doesn't save any
memory because we have to generate a new args list every
function call, and that args list stays in scope durring the
execution of the whole run command because it's not asynchronous
so a simple l-value reference accomplishes the job you wanted
it to do @apachano.